Add cli option for meanies file
Add cli option which allows to give a specific file containig the phrases you
want to use. The phrase that kitty will display will be randomly selected from
the specified file.
The path to the file does not have to be absolute.
In case the `--meanie` option is used at the same time, the `--meanies` option
will be ignored.
 - `--meanies`: type string, default=None

 Usage:
 $ cateow --meanies path/to/file